src/test/java/com/beardream/service/UserServiceTest.java
package com.beardream.service;

import com.beardream.model.User;
import org.junit.Assert;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;

import java.util.List;

import static org.junit.Assert.*;

/**
 * Created by soft01 on 2017/4/18.
 */
@RunWith(SpringRunner.class)
@SpringBootTest
public class UserServiceTest {

    @Autowired
    private UserService userService;

    @Test
    public void userList() throws Exception {
        List<User> list = userService.userList();
        Assert.assertEquals("张弛",list.get(0).getUsername());
    }

}